requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
deepface==0.0.89
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.3
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.4.0
gdown==5.1.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.1
gunicorn==21.2.0
h5py==3.10.0
idna==3.6
importlib_metadata==7.1.0
install==1.3.5
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jax==0.4.26
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==2.12.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mtcnn==0.1.1
namex==0.0.7
nest_asyncio==1.6.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.3.0
pip==23.3.1
platformdirs==4.2.0
prompt-toolkit==3.0.42
protobuf==4.25.3
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0
pytz==2024.1
pywin32==305.1
pyzmq==25.1.2
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
retina-face==0.0.16
rich==13.7.1
rsa==4.9
scipy==1.13.0
setuptools==68.2.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.2
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf_keras==2.16.0
tk==0.1.0
tornado==6.2
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
wheel==0.41.2
wrapt==1.14.1
zipp==3.17.0
